# doc/api.yaml
openapi: 3.0.4
info:
  title: LoginGO API
  description: An API to organize a login template
  version: 1.0.9


servers:
  - url: http://localhost:8080/

paths:

  /:
    get:
      summary: Public landing page
      description: Serves the main homepage for non-authenticated users
      responses:
        "200":
          description: Public homepage
          content:
            text/html:
              schema:
                type: string

  /home:
    get:
      summary: Returns the home page for the logged-in user
      description: Serves the home page after a successful login or registration
      responses:
        "200":
          description: The HTML home page for the session
          content:
            text/html:
              schema:
                type: string
        "401":
          description: Unauthorized - user is not logged in
        "503":
          description: Service unavailable    

  /register:
    get:
      summary: Returns the HTML template to register
      description: Serves the registration page as HTML   
      responses:
        "200":
          description: The HTML login page
          content:
            text/html:
              schema:
                type: string
        "301":
          description: Moved permanently
        "400":
          description: Bad request
        "404":
          description: Not found
        "503":
          description: Service unavailable

    post:
      summary: Registers the user
      description: Registers a new user and redirects to the home page
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "200":
          description: Newly registered user object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "302":
          description: Redirect to the home page after successful login
        "400":
          description: Invalid credentials
        "503":
          description: Service unavailable

  /login:
    get:
      summary: Returns the HTML login page
      description: Serves the login page as HTML
      responses:
        "200":
          description: The HTML login page
          content:
            text/html:
              schema:
                type: string
        "301":
          description: Moved permanently
        "400":
          description: Bad request
        "404":
          description: Not found
        "503":
          description: Service unavailable

    post:
      summary: Logs in the user
      description: Authenticates the user and redirects to the home page
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "302":
          description: Redirect to the home page after successful login
        "400":
          description: Invalid credentials
        "503":
          description: Service unavailable

  /logout:
    post:
      summary: Logs out the user
      responses:
        "200":
          description: Logout successful (redirect or page)
        "301":
          description: Moved permanently
        "400":
          description: Bad request
        "404":
          description: Not found
        "503":
          description: Service unavailable

security:
  - bearerAuth: []

components:
  schemas:
    User:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    error: 
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
  
  securitySchemes:
      bearerAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT
      